Various cleanups.
"Push failed" message is in a p tag with styling.

Try to show the pushed revisions after the header, not before.

Towards merge support. Check for the existence of MERGE_MSG to say whether we're in a merge or not. Disable buttons for real.

Take out the "Hide help" link since I don't feel like supporting it right now.