Don't save the current section if it has a slash in it.
If you hit /admin and the server runs /admin/blog/post_edit, then your browser
will have an incorrect mental model of where you are in the directory tree.
URLs get messed up. It's a bad trip.

If this is a strongly desired feature, it can be implemented -- when we decide
to re-use the last URL from the user object, check if it has a slash, and if
so, then send a redirect.